RSC: Combine server with StreamingSSR server
