+++ packages/cli/src/commands/serveBothHandler.js
@@ -14,19 +14,10 @@ import { getConfig, getPaths } from '@redwoodjs/project-config'
 export const bothExperimentalServerFileHandler = async () => {
   logExperimentalHeader()
 
-  if (getConfig().experimental?.rsc?.enabled) {
-    logSkippingFastifyWebServer()
-
-    await execa(
-      'node',
-      ['./node_modules/@redwoodjs/vite/dist/runRscFeServer.js'],
-      {
-        cwd: getPaths().base,
-        stdio: 'inherit',
-        shell: true,
-      }
-    )
-  } else if (getConfig().experimental?.streamingSsr?.enabled) {
+  if (
+    getConfig().experimental?.rsc?.enabled ||
+    getConfig().experimental?.streamingSsr?.enabled
+  ) {
     logSkippingFastifyWebServer()
 
     await execa('yarn', ['rw-serve-fe'], {
@@ -47,41 +38,17 @@ export const bothExperimentalServerFileHandler = async () => {
   }
 }
 
-export const bothRscServerHandler = async (argv) => {
+export const bothSsrRscServerHandler = async (argv) => {
   const { apiServerHandler } = await import('./serveApiHandler.js')
 
-  // TODO (RSC) Allow specifying port, socket and apiRootPath
+  // TODO Allow specifying port, socket and apiRootPath
   const apiPromise = apiServerHandler({
     ...argv,
     port: 8911,
     apiRootPath: '/',
   })
 
-  // TODO (RSC) More gracefully handle Ctrl-C
-  const fePromise = execa(
-    'node',
-    ['./node_modules/@redwoodjs/vite/dist/runRscFeServer.js'],
-    {
-      cwd: getPaths().base,
-      stdio: 'inherit',
-      shell: true,
-    }
-  )
-
-  await Promise.all([apiPromise, fePromise])
-}
-
-export const bothSsrServerHandler = async (argv) => {
-  const { apiServerHandler } = await import('./serveApiHandler.js')
-
-  // TODO (STREAMING) Allow specifying port, socket and apiRootPath
-  const apiPromise = apiServerHandler({
-    ...argv,
-    port: 8911,
-    apiRootPath: '/',
-  })
-
-  // TODO (STREAMING) More gracefully handle Ctrl-C
+  // TODO More gracefully handle Ctrl-C
   // Right now you get a big red error box when you kill the process
   const fePromise = execa('yarn', ['rw-serve-fe'], {
     cwd: getPaths().web.base,
